Lab_2/Experimento_2/requirements.txt
numpy
matplotlib
imageio
Pillow
IPython
